changed payment account service: -add function to retrieve account by ID -added editorconfig and renamed some files

+++ XWave.Core/DTOs/Customers/PaymentAccountUsageDto.cs
@@ -5,6 +5,6 @@ public record PaymentAccountUsageDto
     public string Provider { get; init; } = string.Empty;
     public string AccountNumber { get; init; } = string.Empty;
     public decimal TotalSpending { get; init; }
-    public ushort PurchaseCount { get; init; }
+    public int PurchaseCount { get; init; }
     public DateTime? LatestPurchase { get; init; }
 }
